fix: return valid byte slice on csaf channel (#43)
Send a newly created byte slice with the csaf document content on the
channel. Previously the sent byte slice was gained from the method
`Buffer.Bytes()`. However this "slice is valid for use only until the
next buffer modification" and therefore modified with the next buffer
modification..

## Why

The bug caused a data race. If the receiver is not processing the
content from the channel fast enough, the content of the returned byte
slice is overwritten by the next downloaded CSAF.