fix: fix memory leak on failed http requests (#59)
In many error cases the http response body was not closed as the `defer`
statement closing the body was only executed after the error handling.

Thus the underlying connection can't be reused. This caused the creation
of many goroutines which are never discarded, massively increasing the
memory consumption.

A prominent scenario is if a CSAF provider does not provide both sha256
and sha512 checksums.

## References

Jira Ticket: [VTI-405](https://jira.greenbone.net/browse/VTI-405)