change: avoid log spam due to not provided sha sums (#71)
A csaf provider is not required to provide both sha256 and sha512 checksums.
Therefor the log message regarding a not present sha sum is expected and
can cause a lot of log messages. This is therefore more suitable on log
channel `debug`.

## Why

Make the info channel well readable containing without noise.