fix: add missing errors on missing signature or checksum (#74)
This is a newly integrated upstream feature, so we need our own error
handling to those checks to get more than just a logged error as
feedback.

*Note:* These error checks only apply to missing references in the ROLIE
feed, this does not change the error handling for the retrieval and
parsing of the signature and checksum files.

## Why

We want to benefit from the improved error handling and fail the
download on ROLIE feeds which don't provide links to all required files
as per CSAF specification.